requirements.txt
openai-whisper
pynput
numpy
python-dotenv
requests
